Refactored Notification. Added more notifications.
Separated Notifications into 3 categories (so far): League, User, and
Like/Comment notifications.

Notifications now have column named content, which is a string which
contains the information relevant to the notification. It will be
displayed to the user in the notifications list.

Also fixed the regex when a user has an avatar for the notification
list to deal with all image types and not just jpg.